rag_pdf/requirements.txt
streamlit
openai
llama-index
python-dotenv
PyPDF2